requirements.txt
python-telegram-bot
psycopg2
Fabric3==1.12.0.post1
ipython<6.3.0
Jinja2<2.9.7
croniter<0.3.20
mypy